pyproject.toml
[project]
name = "building-applications-with-ai-agents"
version = "0.1.0"
description = "A unified, multi-framework platform for designing, implementing, and evaluating AI-powered agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typing-extensions>=4.10",
    "openai>=1.25",
    "tiktoken>=0.5.1",
    "langchain>=1.1.0",
    "langchain-openai>=1.0.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.2.1",
    "langchain-community>=0.4.1",
    "wikipedia>=1.4.0",
    "bert-score>=0.3.13",
    "sentence-transformers>=2.2.2",
    "traceloop-sdk>=0.17.0",
    "websockets>=15.0.1",
    "fastapi>=0.122.0",
    "uvicorn>=0.38.0",
    "requests>=2.32.5",
    "langchain-mcp-adapters>=0.1.14",
    "rank-bm25>=0.2.2",
    "vectordb2>=0.1.9",
    "tensorflow-hub>=0.16.1",
    "backoff>=2.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

